requirements.txt
black>=24.4.2
pytest>=8.2.2
python-dotenv>=1.0.1

httpx>=0.27.0
fastapi>=0.111.0
uvicorn>=0.30.1

PyJWT>=2.8.0